Fix PadFloat still using decimal part if when it should
Move pads to utils
